requirements.txt
absl-py==2.3.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.6.0
cycler==0.12.1
dill==0.4.0
exceptiongroup==1.3.0
exejs==0.0.4
faster-whisper==1.1.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jh2==5.0.9
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.4.2
niquests==3.14.1
numba==0.61.2
numpy==1.26.4
onnxruntime==1.15.1
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pathos==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pox==0.3.6
ppft==1.7.7
protobuf==4.25.8
pycparser==2.22
pydub==0.25.1
PyMatting==1.1.14
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytube==15.0.0
pytube3==9.6.4
PyYAML==6.0.2
qh3==1.5.3
referencing==0.36.2
regex==2024.11.6
rembg==2.0.67
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SpeechRecognition==3.14.3
sympy==1.14.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
translators==6.0.1
typing_extensions==4.14.1
urllib3==2.5.0
urllib3-future==2.13.900
wassima==1.2.2
Werkzeug==3.1.3
yt-dlp==2025.6.30
